+++ packages/date/types/datepicker/LmDatepicker.d.ts
@@ -1,3 +1,3 @@
 import { LmDatepickerProps } from './datepickerTypes';
-export declare function LmDatepicker({ numberOfMonths, isRangePicker, onChange, labelFunctions, startDate, endDate, required, error, helperText, label, labelProps, labelInline, fullWidth, inputProps, buttonProps, }: LmDatepickerProps): import("react/jsx-runtime").JSX.Element;
+export declare function LmDatepicker({ numberOfMonths, isRangePicker, onChange, labelFunctions, startDate, endDate, required, error, helperText, label, labelProps, labelInline, fullWidth, inputProps, buttonProps, popoverProps, }: LmDatepickerProps): import("react/jsx-runtime").JSX.Element;
 //# sourceMappingURL=LmDatepicker.d.ts.map