+++ packages/date/types/datepicker/datepickerTypes.d.ts
@@ -1,6 +1,7 @@
 import { LmFormContainerBaseTypes } from '@tamagui-extras/form';
 import { OnDatesChangeProps, UseMonthProps } from '@datepicker-react/hooks';
 import { ButtonProps, InputProps } from 'tamagui';
+import { LmPopoverProps } from '@tamagui-extras/core';
 export type LmDatepickerProps = LmFormContainerBaseTypes & {
     startDate?: Date | null;
     endDate?: Date | null;
@@ -12,5 +13,6 @@ export type LmDatepickerProps = LmFormContainerBaseTypes & {
     fullWidth?: boolean;
     inputProps?: InputProps;
     buttonProps?: ButtonProps;
+    popoverProps?: Omit<LmPopoverProps, 'trigger'>;
 };
 //# sourceMappingURL=datepickerTypes.d.ts.map