+++ packages/file/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tamagui-extras/file
 
+## 1.1.5
+
+### Patch Changes
+
+- Update to latest version and forward Popover props on Datepicker component
+
 ## 1.1.4
 
 ### Patch Changes