+++ packages/youtube/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tamagui-extras/youtube
 
+## 1.1.7
+
+### Patch Changes
+
+- Update to latest version and forward Popover props on Datepicker component
+- Updated dependencies
+  - @tamagui-extras/core@1.1.7
+
 ## 1.1.6
 
 ### Patch Changes